+++ src/main/java/fiskfille/tf/client/displayable/DisplayableTransformiumSeed.java
@@ -1,10 +1,5 @@
 package fiskfille.tf.client.displayable;
 
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.renderer.OpenGlHelper;
-import net.minecraft.client.renderer.RenderHelper;
-import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.client.renderer.entity.RenderManager;
 import net.minecraft.client.renderer.tileentity.TileEntityRendererDispatcher;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.MathHelper;
@@ -15,20 +10,20 @@ import fiskfille.tf.common.tileentity.TileEntityTransformiumSeed;
 
 public class DisplayableTransformiumSeed extends Displayable
 {
-	@Override
-	public void render(ItemStack itemstack)
-	{
-		renderTag("x" + itemstack.stackSize, 0, 0.05F, 0);
-		
-		GL11.glPushMatrix();
-		float f = 0.5F;
-		float f1 = MathHelper.sin(((float)mc.thePlayer.ticksExisted) / 15.0F) * 0.07F;
-		GL11.glRotatef((float)mc.thePlayer.ticksExisted * 0.75F, 0.0F, 1.0F, 0.0F);
-		GL11.glTranslatef(0, 0.6F + f1, 0);
-		GL11.glRotatef(180, 1, 0, 0);
-		GL11.glScalef(f, f, f);
-		
-		TileEntityRendererDispatcher.instance.renderTileEntityAt(new TileEntityTransformiumSeed(), -0.5F, -0.5F, -0.5F, 0.0F);
-		GL11.glPopMatrix();
-	}
+    @Override
+    public void render(ItemStack itemstack)
+    {
+        renderTag("x" + itemstack.stackSize, 0, 0.05F, 0);
+
+        GL11.glPushMatrix();
+        float f = 0.5F;
+        float f1 = MathHelper.sin(mc.thePlayer.ticksExisted / 15.0F) * 0.07F;
+        GL11.glRotatef(mc.thePlayer.ticksExisted * 0.75F, 0.0F, 1.0F, 0.0F);
+        GL11.glTranslatef(0, 0.6F + f1, 0);
+        GL11.glRotatef(180, 1, 0, 0);
+        GL11.glScalef(f, f, f);
+
+        TileEntityRendererDispatcher.instance.renderTileEntityAt(new TileEntityTransformiumSeed(), -0.5F, -0.5F, -0.5F, 0.0F);
+        GL11.glPopMatrix();
+    }
 }
