+++ src/main/java/fiskfille/tf/client/displayable/DisplayableVehicle.java
@@ -1,6 +1,5 @@
 package fiskfille.tf.client.displayable;
 
-import net.minecraft.client.Minecraft;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
 
@@ -14,57 +13,57 @@ import fiskfille.tf.common.transformer.base.Transformer;
 
 public class DisplayableVehicle extends Displayable
 {
-	@Override
-	public void render(ItemStack itemstack)
-	{
-		bindTexture(new ResourceLocation(TransformersMod.modid, "textures/models/" + getTextureFromStack(itemstack)));
+    @Override
+    public void render(ItemStack itemstack)
+    {
+        bindTexture(new ResourceLocation(TransformersMod.modid, "textures/models/" + getTextureFromStack(itemstack)));
         ModelVehicleBase vehicle = getModelFromStack(itemstack);
-        
+
         if (vehicle != null)
         {
-        	GL11.glRotatef((float)mc.thePlayer.ticksExisted * 0.5F, 0.0F, 1.0F, 0.0F);
+            GL11.glRotatef(mc.thePlayer.ticksExisted * 0.5F, 0.0F, 1.0F, 0.0F);
             GL11.glTranslatef(0.0F, -0.2F, 0.0F);
-            
+
             float scale = 0.75F;
             GL11.glScalef(scale, scale, scale);
             vehicle.render();
         }
-	}
-	
-	public ModelVehicleBase getModelFromStack(ItemStack displayItem)
+    }
+
+    public ModelVehicleBase getModelFromStack(ItemStack displayItem)
     {
         Transformer transformer = TransformersAPI.getTransformers().get(displayItem.getItemDamage());
-        
+
         if (transformer != null)
         {
             return TFModelRegistry.getModel(transformer).getVehicleModel();
         }
-        
+
         return null;
     }
-    
+
     public String getTextureFromStack(ItemStack displayItem)
     {
         Transformer transformer = TransformersAPI.getTransformers().get(displayItem.getItemDamage());
-        
+
         if (transformer != null)
         {
             String name = transformer.getName().toLowerCase().replaceAll(" ", "_");
             return name + "/" + name + ".png";
         }
-        
+
         return null;
     }
-    
+
     public float getBlockHeight(ItemStack itemstack)
-	{
-    	int i = itemstack.getItemDamage();
-    	
-    	if (i == 2)
-    	{
-    		return 0.85F;
-    	}
-    	
-		return 1.0F;
-	}
+    {
+        int i = itemstack.getItemDamage();
+
+        if (i == 2)
+        {
+            return 0.85F;
+        }
+
+        return 1.0F;
+    }
 }
